broke apart utils module into a sub package and updated references
