+++ appdaemon/scheduler.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import asyncio
 import functools
 import logging
@@ -15,12 +17,16 @@ import pytz
 from astral import LocationInfo
 from astral.location import Location
 
-from . import parse, utils
-from .types import TimeDeltaLike
+from .utils import parse
+from .utils.datetime import SUN_EVENT_INTERVAL, ensure_timezone, now_is_between
+from .utils.functools import _sanitize_kwargs, get_kwargs, resolve_offset, unwrapped, validate_offset_within_interval
+from .utils.str import dt_to_str, format_seconds
+from .utils.threading import run_in_executor
 
 if TYPE_CHECKING:
     from .adbase import ADBase
     from .appdaemon import AppDaemon
+    from .types import TimeDeltaLike
 
 
 time_regex_str = r"(?P<hour>\d+):(?P<minute>\d+):(?P<second>\d+)(?:\.(?P<microsecond>\d+))?"
@@ -126,12 +132,12 @@ class Scheduler:
         # self.active = True
 
         if self.AD.starttime is not None:
-            self.now = utils.ensure_timezone(self.AD.starttime, pytz.utc)
+            self.now = ensure_timezone(self.AD.starttime, pytz.utc)
         else:
             self.now = await self.get_now(pytz.utc)
 
         if self.AD.endtime is not None:
-            self.endtime = utils.ensure_timezone(self.AD.endtime, pytz.utc)
+            self.endtime = ensure_timezone(self.AD.endtime, pytz.utc)
         else:
             self.endtime = None
 
@@ -192,19 +198,19 @@ class Scheduler:
         if repeat:
             match type_:
                 case "next_rising":
-                    utils.validate_offset_within_interval(
-                        offset, utils.SUN_EVENT_INTERVAL, "sunrise", random_start, random_end
+                    validate_offset_within_interval(
+                        offset, SUN_EVENT_INTERVAL, "sunrise", random_start, random_end
                     )
                 case "next_setting":
-                    utils.validate_offset_within_interval(
-                        offset, utils.SUN_EVENT_INTERVAL, "sunset", random_start, random_end
+                    validate_offset_within_interval(
+                        offset, SUN_EVENT_INTERVAL, "sunset", random_start, random_end
                     )
                 case _ if interval.total_seconds() > 0:
-                    utils.validate_offset_within_interval(
+                    validate_offset_within_interval(
                         offset, interval, "interval", random_start, random_end
                     )
 
-        c_offset = utils.resolve_offset(offset=offset, random_start=random_start, random_end=random_end)
+        c_offset = resolve_offset(offset=offset, random_start=random_start, random_end=random_end)
         timestamp = basetime + c_offset
 
         self.schedule[name][handle] = {
@@ -238,8 +244,8 @@ class Scheduler:
             state="active",
             attributes={
                 "app": name,
-                "execution_time": utils.dt_to_str(timestamp, self.AD.tz, round=True),
-                "repeat": str(utils.parse_timedelta(interval)),
+                "execution_time": dt_to_str(timestamp, self.AD.tz, round=True),
+                "repeat": str(parse.parse_timedelta(interval)),
                 "function": function_name,
                 "pinned": pin_app,
                 "pinned_thread": pin_thread,
@@ -292,7 +298,7 @@ class Scheduler:
             case _:
                 raise ValueError("Malformed scheduler args, expected 'type' or 'interval' key")
 
-        c_offset = utils.resolve_offset(
+        c_offset = resolve_offset(
             offset=args.get("offset"),
             random_start=args.get("random_start"),
             random_end=args.get("random_end"),
@@ -300,7 +306,7 @@ class Scheduler:
         args["timestamp"] = args["basetime"] + c_offset
 
         # Update entity
-        execution_time = utils.dt_to_str(args["timestamp"].replace(microsecond=0), self.AD.tz)
+        execution_time = dt_to_str(args["timestamp"].replace(microsecond=0), self.AD.tz)
         await self.AD.state.set_state(
             "_scheduler",
             "admin",
@@ -319,7 +325,7 @@ class Scheduler:
             )  # fmt: skip
             return False
 
-        args = await utils.run_in_executor(self, deepcopy, self.schedule[name][handle])
+        args = await run_in_executor(self, deepcopy, self.schedule[name][handle])
         match args:
             case {"type": "next_rising" | "next_setting"}:
                 self.logger.warning(
@@ -359,7 +365,7 @@ class Scheduler:
                 "basetime": datetime() as basetime,
                 "interval": timedelta() as interval,
             }:
-                callback_name = utils.unwrapped(callback).__name__
+                callback_name = unwrapped(callback).__name__
                 logger.debug(f"callback name={callback_name}")
                 logger.debug(f"     basetime={basetime.astimezone(self.AD.tz).isoformat()}")
                 logger.debug(f"    timestamp={timestamp.astimezone(self.AD.tz).isoformat()}")
@@ -490,7 +496,7 @@ class Scheduler:
         Otherwise, calculates a start time (defaulting to "now") and advances by the
         interval until a future time is reached.
         """
-        interval = utils.parse_timedelta(interval)
+        interval = parse.parse_timedelta(interval)
         start = "now" if start is None else start
 
         # Get "now" once and use it consistently to avoid timing races
@@ -608,7 +614,7 @@ class Scheduler:
                             # No kick, no scheduler expiry ...
                             delta = idle_time
 
-                    self.now += utils.parse_timedelta(delta)
+                    self.now += parse.parse_timedelta(delta)
 
                 self.last_fired = await self.get_now(pytz.utc)
                 self.logger.debug("self.now   utc=%s", self.last_fired.isoformat())
@@ -652,7 +658,7 @@ class Scheduler:
                             time_to_run = timestamp <= self.now
                             args = self.schedule.get(name, {}).get(uuid_, False)
                             if time_to_run and args:
-                                func = utils.unwrapped(args["callback"])
+                                func = unwrapped(args["callback"])
                                 if func is not None:
                                     self.logger.debug("Firing scheduled callback %s for '%s'", func.__name__, name)
                                 else:
@@ -681,7 +687,7 @@ class Scheduler:
                 # sleep in and potentially miss an event that should happen earlier than expected due to the time change
                 #
 
-                next = self.now + utils.parse_timedelta(delay)
+                next = self.now + parse.parse_timedelta(delay)
 
                 self.logger.debug("next event=%s", next.astimezone(self.AD.tz).isoformat())
 
@@ -770,18 +776,18 @@ class Scheduler:
                 schedule[name][str(entry)]["basetime"] = str(self.AD.sched.make_naive(self.schedule[name][entry]["basetime"]))
                 schedule[name][str(entry)]["repeat"] = self.schedule[name][entry]["repeat"]
                 if self.schedule[name][entry]["type"] == "next_rising":
-                    schedule[name][str(entry)]["interval"] = "sunrise:{}".format(utils.format_seconds(self.schedule[name][entry]["offset"]))
+                    schedule[name][str(entry)]["interval"] = "sunrise:{}".format(format_seconds(self.schedule[name][entry]["offset"]))
                 elif self.schedule[name][entry]["type"] == "next_setting":
-                    schedule[name][str(entry)]["interval"] = "sunset:{}".format(utils.format_seconds(self.schedule[name][entry]["offset"]))
+                    schedule[name][str(entry)]["interval"] = "sunset:{}".format(format_seconds(self.schedule[name][entry]["offset"]))
                 elif self.schedule[name][entry]["repeat"] is True:
-                    schedule[name][str(entry)]["interval"] = utils.format_seconds(self.schedule[name][entry]["interval"])
+                    schedule[name][str(entry)]["interval"] = format_seconds(self.schedule[name][entry]["interval"])
                 else:
                     schedule[name][str(entry)]["interval"] = "None"
 
                 schedule[name][str(entry)]["offset"] = self.schedule[name][entry]["offset"]
                 schedule[name][str(entry)]["kwargs"] = ""
                 for kwarg in self.schedule[name][entry]["kwargs"]:
-                    schedule[name][str(entry)]["kwargs"] = utils.get_kwargs(self.schedule[name][entry]["kwargs"])
+                    schedule[name][str(entry)]["kwargs"] = get_kwargs(self.schedule[name][entry]["kwargs"])
                 if self.schedule[name][entry]["callback"] is None:
                     schedule[name][str(entry)]["callback"] = "cancel_callback"
                 elif isinstance(self.schedule[name][entry]["callback"], functools.partial):
@@ -845,7 +851,7 @@ class Scheduler:
         now: datetime | None = None,
     ) -> bool:
         now = now if now is not None else await self.get_now()
-        return utils.now_is_between(
+        return now_is_between(
             now=now.astimezone(self.AD.tz),  # Need to force timezone during time-travel mode
             start_time=start_time,
             end_time=end_time,
@@ -912,7 +918,7 @@ class Scheduler:
         # Need to force timezone during time-travel mode
         if now is None:
             now = await self.get_now()
-        now = utils.ensure_timezone(now, self.AD.tz)
+        now = ensure_timezone(now, self.AD.tz)
         return parse.parse_datetime(
             input_=input_,
             now=now,
@@ -967,7 +973,7 @@ class Scheduler:
     def sanitize_timer_kwargs(app: "ADBase", kwargs: dict) -> dict:
         """Removes keywords from the keywords"""
         kwargs_copy = kwargs.copy()
-        return utils._sanitize_kwargs(
+        return _sanitize_kwargs(
             kwargs_copy,
             ["interval", "constrain_days", "constrain_input_boolean", "_pin_app", "_pin_thread", "__silent"] + app.constraints,
         )
