+++ appdaemon/threads.py
@@ -17,8 +17,14 @@ from typing import TYPE_CHECKING, Any, ClassVar
 import iso8601
 
 from . import exceptions as ade
-from . import utils
 from .models.config.app import AppConfig
+from .utils import parse
+from .utils.datetime import day_of_week
+from .utils.functools import has_expanded_kwargs
+from .utils.misc import deepcopy as ad_deepcopy
+from .utils.state import check_state
+from .utils.str import dt_to_str, format_timedelta, str_to_dt
+from .utils.threading import run_async_sync_func, run_coroutine_threadsafe, run_in_executor
 
 if TYPE_CHECKING:
     from .adbase import ADBase
@@ -185,12 +191,12 @@ class Threading:
         await self.add_entity(
             "admin",
             "sensor.threads_max_busy_time",
-            utils.dt_to_str(datetime.datetime(1970, 1, 1, 0, 0, 0, 0)),
+            dt_to_str(datetime.datetime(1970, 1, 1, 0, 0, 0, 0)),
         )
         await self.add_entity(
             "admin",
             "sensor.threads_last_action_time",
-            utils.dt_to_str(datetime.datetime(1970, 1, 1, 0, 0, 0, 0)),
+            dt_to_str(datetime.datetime(1970, 1, 1, 0, 0, 0, 0)),
         )
 
     async def create_initial_threads(self) -> None:
@@ -241,7 +247,7 @@ class Threading:
             {
                 "q": 0,
                 "is_alive": True,
-                "time_called": utils.dt_to_str(datetime.datetime(1970, 1, 1, 0, 0, 0, 0)),
+                "time_called": dt_to_str(datetime.datetime(1970, 1, 1, 0, 0, 0, 0)),
                 "pinned_apps": [],
             },
         )
@@ -297,7 +303,7 @@ class Threading:
         self.diag.info("--------------------------------------------------")
         current_busy = await self.get_state("_threading", "admin", "sensor.threads_current_busy")
         max_busy = await self.get_state("_threading", "admin", "sensor.threads_max_busy")
-        max_busy_time = utils.str_to_dt(await self.get_state("_threading", "admin", "sensor.threads_max_busy_time"))
+        max_busy_time = str_to_dt(await self.get_state("_threading", "admin", "sensor.threads_max_busy_time"))
         last_action_time = await self.get_state("_threading", "admin", "sensor.threads_last_action_time")
         self.diag.info("Currently busy threads: %s", current_busy)
         self.diag.info("Most used threads: %s at %s", max_busy, max_busy_time)
@@ -372,7 +378,7 @@ class Threading:
                     id = thread_id.split("-")[1]
                     await self.add_thread(silent=False, id=id)
                 if await self.get_state("_threading", "admin", "thread.{}".format(thread_id)) != "idle":
-                    start = utils.str_to_dt(
+                    start = str_to_dt(
                         await self.get_state(
                             "_threading",
                             "admin",
@@ -450,7 +456,7 @@ class Threading:
 
         now = await self.AD.sched.get_now()
         if callback == "idle":
-            start = utils.str_to_dt(
+            start = str_to_dt(
                 await self.get_state(
                     "_threading",
                     "admin",
@@ -468,8 +474,8 @@ class Threading:
                 callback = await self.get_state("_threading", "admin", thread_name)
                 self.logger.warning(
                     f"Excessive time spent in callback {callback}. "
-                    f"Thread entity: '{thread_name}' - now complete after {utils.format_timedelta(duration)} "
-                    f"(limit={utils.format_timedelta(self.AD.thread_duration_warning_threshold)})"
+                    f"Thread entity: '{thread_name}' - now complete after {format_timedelta(duration)} "
+                    f"(limit={format_timedelta(self.AD.thread_duration_warning_threshold)})"
                 )
             await self.add_to_state("_threading", "admin", "sensor.threads_current_busy", -1)
 
@@ -497,14 +503,14 @@ class Threading:
                 "_threading",
                 "admin",
                 "sensor.threads_max_busy_time",
-                state=utils.dt_to_str((await self.AD.sched.get_now()).replace(microsecond=0), self.AD.tz),
+                state=dt_to_str((await self.AD.sched.get_now()).replace(microsecond=0), self.AD.tz),
             )
 
             await self.set_state(
                 "_threading",
                 "admin",
                 "sensor.threads_last_action_time",
-                state=utils.dt_to_str((await self.AD.sched.get_now()).replace(microsecond=0), self.AD.tz),
+                state=dt_to_str((await self.AD.sched.get_now()).replace(microsecond=0), self.AD.tz),
             )
 
         # Update thread info
@@ -516,7 +522,7 @@ class Threading:
                 "thread.{}".format(thread_id),
                 q=0,
                 state=callback,
-                time_called=utils.dt_to_str(now.replace(microsecond=0), self.AD.tz),
+                time_called=dt_to_str(now.replace(microsecond=0), self.AD.tz),
                 is_alive=True,
                 pinned_apps=[],
             )
@@ -527,7 +533,7 @@ class Threading:
                 "thread.{}".format(thread_id),
                 q=self.threads[thread_id]["queue"].qsize(),
                 state=callback,
-                time_called=utils.dt_to_str(now.replace(microsecond=0), self.AD.tz),
+                time_called=dt_to_str(now.replace(microsecond=0), self.AD.tz),
                 is_alive=self.threads[thread_id]["thread"].is_alive(),
                 pinned_apps=self.get_pinned_apps(thread_id),
             )
@@ -551,7 +557,7 @@ class Threading:
                 "admin",
                 thread_entity,
                 "idle",
-                {"q": 0, "is_alive": True, "time_called": utils.dt_to_str(datetime.datetime(1970, 1, 1, 0, 0, 0, 0))},
+                {"q": 0, "is_alive": True, "time_called": dt_to_str(datetime.datetime(1970, 1, 1, 0, 0, 0, 0))},
             )
             self.threads[thread.name] = {}
             self.threads[thread.name]["queue"] = Queue(maxsize=0)
@@ -656,7 +662,7 @@ class Threading:
         unconstrained = True
         if hasattr(app, "constraints") and key in app.constraints:
             method = getattr(app, key)
-            unconstrained = await utils.run_async_sync_func(self, method, value)
+            unconstrained = await run_async_sync_func(self, method, value)
 
         return unconstrained
 
@@ -688,7 +694,7 @@ class Threading:
             now = (await self.AD.sched.get_now()).astimezone(self.AD.tz)
             daylist = []
             for day in days.split(","):
-                daylist.append(await utils.run_in_executor(self, utils.day_of_week, day))
+                daylist.append(await run_in_executor(self, day_of_week, day))
 
             if now.weekday() not in daylist:
                 unconstrained = False
@@ -700,7 +706,7 @@ class Threading:
 
         unconstrained = True
         if "constrain_state" in args:
-            unconstrained = utils.check_state(self.logger, new_state, args["constrain_state"], name)
+            unconstrained = check_state(self.logger, new_state, args["constrain_state"], name)
 
         return unconstrained
 
@@ -802,7 +808,7 @@ class Threading:
                         #
                         # Set a timer
                         #
-                        exec_time = await self.AD.sched.get_now() + utils.parse_timedelta(kwargs["duration"])
+                        exec_time = await self.AD.sched.get_now() + parse.parse_timedelta(kwargs["duration"])
 
                         #
                         # If it's a oneshot, scheduler will delete the callback once it has executed,
@@ -881,7 +887,7 @@ class Threading:
         #
         # Callback level constraints
         #
-        myargs = utils.deepcopy(args)
+        myargs = ad_deepcopy(args)
         if "kwargs" in myargs:
             for arg in myargs["kwargs"].keys():
                 constrained = await self.check_constraint(
@@ -979,7 +985,7 @@ class Threading:
                         pos_args = (args["event"], data)
                         kwargs = self.AD.events.sanitize_event_kwargs(app, args["kwargs"])
 
-                use_dictionary_unpacking = utils.has_expanded_kwargs(funcref)
+                use_dictionary_unpacking = has_expanded_kwargs(funcref)
                 if use_dictionary_unpacking:
                     funcref = functools.partial(funcref, *pos_args, **kwargs)
                 else:
@@ -1071,7 +1077,7 @@ class Threading:
                             pos_args = (args["event"], args["data"])
                             kwargs = self.AD.events.sanitize_event_kwargs(app, args["kwargs"])
 
-                    use_dictionary_unpacking = utils.has_expanded_kwargs(funcref)
+                    use_dictionary_unpacking = has_expanded_kwargs(funcref)
                     if use_dictionary_unpacking:
                         funcref = functools.partial(funcref, *pos_args, **kwargs)
                     else:
@@ -1084,7 +1090,7 @@ class Threading:
 
                     callback = f"{funcref.func.__qualname__} for {name}"
                     update_coro = self.update_thread_info(thread_id, callback, name, _type, _id, silent)
-                    utils.run_coroutine_threadsafe(self, update_coro)
+                    run_coroutine_threadsafe(self, update_coro)
 
                     @ade.wrap_sync(error_logger, self.AD.app_dir, callback)
                     def safe_callback():
@@ -1110,7 +1116,7 @@ class Threading:
 
                 finally:
                     update_coro = self.update_thread_info(thread_id, "idle", name, _type, _id, silent)
-                    utils.run_coroutine_threadsafe(self, update_coro)
+                    run_coroutine_threadsafe(self, update_coro)
                     q.task_done()  # Have this in multiple places to ensure it gets called even if an exception is raised
             else:
                 if not self.AD.stopping:
@@ -1146,7 +1152,7 @@ class Threading:
             "terminate": {"count": 0, "signature": {True: "terminate()", False: "terminate()"}},
         }
 
-        use_dictionary_unpacking = utils.has_expanded_kwargs(funcref)
+        use_dictionary_unpacking = has_expanded_kwargs(funcref)
 
         try:
             if isinstance(funcref, functools.partial):
