+++ appdaemon/utils/perf.py
@@ -1,0 +1,45 @@
+import cProfile
+import io
+import pstats
+import sys
+
+
+def _profile_this(fn):
+    def profiled_fn(*args, **kwargs):
+        self = args[0]
+        self.pr = cProfile.Profile()
+        self.pr.enable()
+
+        result = fn(self, *args, **kwargs)
+
+        self.pr.disable()
+        s = io.StringIO()
+        sortby = "cumulative"
+        ps = pstats.Stats(self.pr, stream=s).sort_stats(sortby)
+        ps.print_stats()
+        self.profile = fn + s.getvalue()
+
+        return result
+
+    return profiled_fn
+
+
+def get_object_size(obj, seen=None):
+    """Recursively finds size of objects"""
+    size = sys.getsizeof(obj)
+    if seen is None:
+        seen = set()
+    obj_id = id(obj)
+    if obj_id in seen:
+        return 0
+    # Important mark as seen *before* entering recursion to gracefully handle
+    # self-referential objects
+    seen.add(obj_id)
+    if isinstance(obj, dict):
+        size += sum([get_object_size(v, seen) for v in obj.values()])
+        size += sum([get_object_size(k, seen) for k in obj.keys()])
+    elif hasattr(obj, "__dict__"):
+        size += get_object_size(obj.__dict__, seen)
+    elif hasattr(obj, "__iter__") and not isinstance(obj, (str, bytes, bytearray)):
+        size += sum([get_object_size(i, seen) for i in obj])
+    return size
