+++ tests/unit/datetime/test_datetime_misc.py
@@ -1,7 +1,8 @@
 from datetime import datetime, timedelta
 
 import pytest
-from appdaemon import utils
+from appdaemon.utils.datetime import ensure_timezone
+from appdaemon.utils.functools import resolve_offset
 from pytz import BaseTzInfo
 
 pytestmark = [
@@ -12,7 +13,7 @@ pytestmark = [
 
 def test_resolve_offset() -> None:
     offsets = sorted(
-        utils.resolve_offset(
+        resolve_offset(
             timedelta(seconds=10),
             random_start=timedelta(seconds=-5),
             random_end=timedelta(seconds=5)
@@ -26,7 +27,7 @@ def test_resolve_offset() -> None:
 
 def test_ensure_timezone(tz: BaseTzInfo) -> None:
     naive = datetime(2025, 6, 25, 12, 0, 0)
-    aware = utils.ensure_timezone(naive, tz)
+    aware = ensure_timezone(naive, tz)
     assert aware.tzinfo is not None, "Datetime should be timezone-aware"
     assert naive.time() == aware.time(), "Time should remain the same after ensuring timezone"
     assert aware != datetime(2025, 6, 25, 12, 0, 0, tzinfo=tz)
