+++ tests/unit/datetime/test_now_is_between.py
@@ -2,7 +2,7 @@ from datetime import datetime, time
 from functools import partial
 
 import pytest
-from appdaemon import utils
+from appdaemon.utils.datetime import now_is_between
 from astral.location import Location
 
 pytestmark = [
@@ -12,7 +12,7 @@ pytestmark = [
 
 
 def test_between_overnight(location: Location, early_now: datetime, default_now: datetime, late_now: datetime) -> None:
-    sun_check = partial(utils.now_is_between, start_time="sunset", end_time="sunrise", location=location)
+    sun_check = partial(now_is_between, start_time="sunset", end_time="sunrise", location=location)
     assert sun_check(now=early_now), "The early time is not between sunset and sunrise, but should be"
     assert not sun_check(now=default_now), "The default time is between sunset and sunrise, but should not be"
     assert sun_check(now=late_now), "The late time is not between sunset and sunrise, but should be"
@@ -20,7 +20,7 @@ def test_between_overnight(location: Location, early_now: datetime, default_now:
     # Test again with some offsets
     offset = "03:00:00"
     sun_check = partial(
-        utils.now_is_between,
+        now_is_between,
         start_time="sunset",
         end_time=f"sunrise - {offset}",
         location=location
@@ -30,7 +30,7 @@ def test_between_overnight(location: Location, early_now: datetime, default_now:
     assert sun_check(now=late_now), "The late time is not between sunset and the offset sunrise, but should be"
 
     sun_check = partial(
-        utils.now_is_between,
+        now_is_between,
         start_time=f"sunset + {offset}",
         end_time="sunrise",
         location=location
@@ -41,31 +41,31 @@ def test_between_overnight(location: Location, early_now: datetime, default_now:
 
 
 def test_between_times(location: Location, early_now: datetime, default_now: datetime, late_now: datetime) -> None:
-    between_check = partial(utils.now_is_between, start_time="22:00:00", end_time="02:00:00", location=location)
+    between_check = partial(now_is_between, start_time="22:00:00", end_time="02:00:00", location=location)
     assert not between_check(now=early_now)
     assert not between_check(now=default_now)
     assert between_check(now=late_now)
 
-    between_check = partial(utils.now_is_between, start_time="22:00:00", end_time="08:00:00", location=location)
+    between_check = partial(now_is_between, start_time="22:00:00", end_time="08:00:00", location=location)
     assert between_check(now=early_now)
     assert not between_check(now=default_now)
     assert between_check(now=late_now)
 
     # Should work with time objects
-    between_check = partial(utils.now_is_between, start_time=time(22, 0, 0), end_time=time(8, 0, 0), location=location)
+    between_check = partial(now_is_between, start_time=time(22, 0, 0), end_time=time(8, 0, 0), location=location)
     assert between_check(now=early_now)
     assert not between_check(now=default_now)
     assert between_check(now=late_now)
 
     # Should work with datetime time objects
-    between_check = partial(utils.now_is_between, start_time=default_now.replace(hour=22), end_time=default_now.replace(hour=8), location=location)
+    between_check = partial(now_is_between, start_time=default_now.replace(hour=22), end_time=default_now.replace(hour=8), location=location)
     assert between_check(now=early_now)
     assert not between_check(now=default_now)
     assert between_check(now=late_now)
 
 
 def test_between_simple(default_now: datetime, location: Location) -> None:
-    check = partial(utils.now_is_between, now=default_now, location=location)
+    check = partial(now_is_between, now=default_now, location=location)
     assert not check(start_time="10:00:00", end_time="11:00:00"), "Both times before 'now'"
     assert check(start_time="11:00:00", end_time="13:00:00"), "'Now' is between the two times"
     assert not check(start_time="13:00:00", end_time="14:00:00"), "Both times are after 'now'"
